\name{binarizationExample}
\alias{binarizationExample}
\docType{data}
\title{
An artificial data set consisting of ten artificial feature vectors.
}
\description{
An artificial data set consisting of ten artificial feature vectors that are used to illustrate the binarization methods in the package vignette. Each row of the matrix \code{binarizationExample} corresponds to one feature vector, of which 10 measurements are drawn from a normal distribution N(0,1). The remaining 10 measurements are drawn from a normal distribution N(m,1), with \code{m=10:1} decreasing from the first to the last row.
}
\usage{data(binarizationExample)}
\format{
 The data is a matrix with 20 columns and 10 rows.
}
\keyword{datasets}
