\name{trinarizationExample}
\alias{trinarizationExample}
\docType{data}
\title{
An artificial data set consisting of ten artificial feature vectors.
}
\description{
An artificial data set consisting of 100 artificial feature vectors that are used to illustrate the trinarization methods in the package vignette. Each row of the matrix \code{trinarizationExample} corresponds to one feature vector, of which 5 measurements are drawn from a normal distribution N(0,1). The remaining 10 measurements are drawn from two normal distributions N(m,1), with \code{m=10:1} and \code{m=seq(20,2,by=-2)} (5 meansurements per distribution).
}
\usage{data(trinarizationExample)}
\format{
The data is a matrix with 15 columns and 100 rows.
}
\keyword{datasets}
